moved AnalyticsUtils into new package org.wordpress.android.util.analytics
